bdm/ingestion/streaming/finazon/requirements.txt
click==8.1.8
kafka-python==2.1.1
websockets==15.0.1
